Refactor autoscaler class and combin with update_nodes.py
